requirements.txt
yfinance
pandas
numpy
matplotlib
ta
textblob
newspaper3k
openai
beautifulsoup4
streamlit
scikit-learn
